Galleria_V
Galleria_V is a modification of Aino's JavaScript Image Gallery
"Galleria" and the "Classic" theme. Galleria_V allows for a vertical
carousel and enables fullscreen viewing capability. Galleria_V is not
compatible with Galleria themes, however themes could be modified or
created to function with Galleria_V. Galleria_V retains most of the
functionality of Galleria but is not provided or supported by Aino.